test: use the custom circuit for integration test
